完成查看需求帖子的基本功能；测试修正了编辑需求/项目功能；修改model构造器的不妥之处；修改mapper的resultType；删除不必要的ftl文件；需改进：帖子展示数据库加载逻辑；需求/项目分页；明日todo：列表搜索、点赞/评论功能

+++ src/main/java/cn/edu/nju/rm/dao/RequirementMapper.java
@@ -60,5 +60,5 @@ public interface RequirementMapper {
      * @param pid 项目id
      * @return 已收集需求数
      */
-    int checkNumber (Integer pid);
+    long checkNumber (Integer pid);
 }
